Praca Domowa1

+++ 02-TypesAndVariables/zad3cz.py
@@ -2,6 +2,6 @@ father=5000
 mother=7000
 numofpeople=3
 totalincome=father+mother
-int(incperperson=totalincome/numofpeople)
+incperperson=totalincome/numofpeople
 s=f"Enter father's income: {father}\n Enter mother's income: {mother} \n Enter number of people: {numofpeople}\n Total income: {totalincome}\n Income per person: {incperperson}\n "
 print(s)
